+++ src/main/java/com/example/task/docs/v1/notice/dto/NoticeReqDto.java
@@ -1,4 +1,4 @@
-package com.example.task.docs.notice.dto;
+package com.example.task.docs.v1.notice.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
